Just a beginning
This is one of the first html files I practiced with. I know it doesn't have much but it was just the beginning. I want to show off simple html pages to show my efficiency in html5 and css. 

Also, I just chose this topic for something fun to work with.:)